+++ NEWS.md
@@ -23,6 +23,10 @@ Version 0.15a0 - dev
 - NEW: `matplotlib.qsave()`, `ltype` argument to switch between matplotlib dpi 
   based linetype rendering and AutoCAD like drawing units based linetype 
   rendering
+- NEW: `Solid.vertices()` returns OCS vertices in correct order (also `Trace`)
+- NEW: `Solid.wcs_vertices()` returns WCS vertices in correct order (also `Trace`)
+- NEW: `Face3D.wcs_vertices()` returns WCS vertices, compatibility interface to 
+  SOLID and TRACE 
 - REMOVED: deprecated `DXFEntity.transform_to_wcs()` interface, 
   use `DXFEntity.transform(ucs.matrix)`
 - REMOVED: deprecated `Hatch.edit_boundary()` context manager,